+++ src/main/java/io/jenkins/plugins/appcenter/model/AppCenterServiceFactory.java
@@ -1,4 +1,4 @@
-package io.jenkins.plugins.appcenter.remote;
+package io.jenkins.plugins.appcenter.model;
 
 import hudson.util.Secret;
 import okhttp3.Headers;
