+++ src/main/java/io/jenkins/plugins/appcenter/model/BuildInfo.java
@@ -1,4 +1,4 @@
-package io.jenkins.plugins.appcenter.remote;
+package io.jenkins.plugins.appcenter.model;
 
 public class BuildInfo {
     public final String branch;
