+++ src/main/java/io/jenkins/plugins/appcenter/model/ReleaseDetailsUpdateRequest.java
@@ -1,4 +1,4 @@
-package io.jenkins.plugins.appcenter.remote;
+package io.jenkins.plugins.appcenter.model;
 
 import java.util.List;
