+++ src/main/java/io/jenkins/plugins/appcenter/model/UploadService.java
@@ -1,4 +1,4 @@
-package io.jenkins.plugins.appcenter.remote;
+package io.jenkins.plugins.appcenter.model;
 
 import okhttp3.MultipartBody;
 import retrofit2.http.Multipart;
