+++ src/main/java/io/jenkins/plugins/appcenter/task/UploadTask.java
@@ -3,14 +3,14 @@ package io.jenkins.plugins.appcenter.task;
 import hudson.FilePath;
 import hudson.model.TaskListener;
 import io.jenkins.plugins.appcenter.AppCenterException;
-import io.jenkins.plugins.appcenter.remote.AppCenterServiceFactory;
-import io.jenkins.plugins.appcenter.remote.DestinationId;
-import io.jenkins.plugins.appcenter.remote.ReleaseDetailsUpdateRequest;
-import io.jenkins.plugins.appcenter.remote.ReleaseDetailsUpdateResponse;
-import io.jenkins.plugins.appcenter.remote.ReleaseUploadBeginResponse;
-import io.jenkins.plugins.appcenter.remote.ReleaseUploadEndRequest;
-import io.jenkins.plugins.appcenter.remote.ReleaseUploadEndResponse;
-import io.jenkins.plugins.appcenter.remote.Status;
+import io.jenkins.plugins.appcenter.model.AppCenterServiceFactory;
+import io.jenkins.plugins.appcenter.model.DestinationId;
+import io.jenkins.plugins.appcenter.model.ReleaseDetailsUpdateRequest;
+import io.jenkins.plugins.appcenter.model.ReleaseDetailsUpdateResponse;
+import io.jenkins.plugins.appcenter.model.ReleaseUploadBeginResponse;
+import io.jenkins.plugins.appcenter.model.ReleaseUploadEndRequest;
+import io.jenkins.plugins.appcenter.model.ReleaseUploadEndResponse;
+import io.jenkins.plugins.appcenter.model.Status;
 import okhttp3.MediaType;
 import okhttp3.MultipartBody;
 import okhttp3.RequestBody;
